alx-project-nexus/requirements.txt
Django~=5.2.4
django-environ~=0.12.0
django-cors-headers
djangorestframework~=3.16.0
drf-yasg
psycopg2
django-filter
django-redis
django-cacheops
djangorestframework-simplejwt
httpagentparser~=1.9.5
redis~=6.2.0
celery~=5.5.3
django-mptt~=0.17.0
openai~=1.97.1
django-celery-beat


# For development
django-debug-toolbar~=6.0.0
pytest
